First attempt at metropolis XY model

+++ Assignment_4/include/properties.h
@@ -3,7 +3,7 @@
 
 
 
-double compute_energy(std::vector<std::vector<int> >& lattice);
+double compute_energy(std::vector<std::vector<double> >& lattice);
 void write_properties(std::vector<double> energies, std::string outfile);
 void write_energies(std::vector<double> energies, std::string outfile);
-void write_configuration(std::vector<std::vector<int> >& lattice, std::string outfile);
+void write_configuration(std::vector<std::vector<double> >& lattice, std::string outfile);
